fix speed regression introduced in refactoring
The speed regression (~ 2x) was due to using maps instead of structs or
lists.

During this endeavour I also encountered a data race issue. It was due
to a []string variable that was read and write concurrently. It was
solved by initializing the variable inside the for-loop, instead of
initializing outside the for-loop and reusing it.