clean up TODO comments
* TSV reading/writing todos

For now I don't see a way to have all of:
- mapping of TSV rows to structs
- dynamically set the TSV columns (from conf)
- low complexity

As one way to do this would be to define a struct with tags for each
field, but then this would need using the 'reflect' package, which 
increases the code complexity.

* Stream multiplexing
I think Go doesn't lend itself to this kind of generic refactoring. At
least I am not confident I will manage to refactor this while keeping
the complexity low.