add testing
Problem: there is currently no automatic tests for mmpio, making it
difficult to find regression, trust the system, do refactoring.

What I am about to do: think what kind of tests I want (i.e. what will
the tests cover), implement them, make sure they are automatic and
their status is in a tight feedback loop.

What I expect the result to be: implemented a couple of tests (not
covering the whole codebase, but at least enough so that for the next
change I can write a test with the same system); integrated in GitHub
actions with clear status.


* 2024-02-02  Let's start simple
What would be the simplest test?
Maybe one where we have 2 rows as input, and just get 1 in the output
with the p-value < 1e-6.

I could use go embeds to have input & output files as plain text files,
but make easier (?), slicker (?) to deal with in Go code.
Not sure if good long-term decision, but looks fun! Hype!

OK, starting with making a dummy test to understand how `go test` work
and get going.

Seems `go test` is looking for *_test.go files.

Err, seems there is some (bad) interactions between the flags package
and testing?

By injecting a `fmt.Println(os.Args)` I can see that the program is
called with:
.../mmpio.test -test.paniconexit0 -test.timeout=10m0s

Removing the custom Usage function doesn't remove the error.

Read about the interaction between init() and flag and testing at:
https://go.dev/doc/go1.13#testingpkgtesting

Fixed go test run bug by changing my code to init the flag in main()
instead of init().

Now basic dummy test is working as expected.


* 2024-02-06
Checking status. I should be able to run `go test` and have the dummy
test PASS or FAIL.
Working as expected.

Now turning that into a real test.
Making the dummy data first.

Hmm, sum stats are expected to be gzip-compressed.
Let's see.

Arh. I don't want to put gzipped data files in Git.
It will just make it harder to simply read and write.

What are my options:
1. embed the tsv file, then gzip it during test
2. refactor the main mmpio funcs to allow a compression argument
3. refactor to specify the compression in config file

Option (1.) seems the best since we won't benefit from the refactoring
options elsewhere.
But is it possible? Our current process is that we provide a filepath,
not a gzip buffer.

Can we have a 4th option where we bypass the conf?
We could make a conf struct in the test.
But still need to provide a filepath as `string` to streamTsv.

A 5th option would be to use something else than go test, just bash.
Maybe the best option for end-to-end test.
While `go test` would be good for unit tests.

Let's see how this 5th option would work.
Would need to access an mmpio binary?
Or just with go run maybe.
"The exit status of Run is not the exit status of the compiled binary."
Hmm.

Otherwise I can add a build step in the test.sh, but then how to turn
the test TSV into a GZIP and inject the path somehow in the conf file?

So we can just do option 5, and provide the gzip path in the conf and
make this gzip file on test run.

GH actions fails because the test action doesn't have a build step
so our test script cannot find a ./mmpio to run.

Test working as expected on GH actions! :tada:


Now onto second test, and refactor a little the common bash bits.